:sparkles: Add parser for json... Set, Get and Del commands work successfully

+++ value.c
@@ -8,7 +8,7 @@ JsonValue* allocNumber(long long num) {
   return value;
 }
 
-JsonValue* allocObject(uint64_t size) {
+JsonValue* allocObject(size_t size) {
   JsonValue* value = RedisModule_Calloc(1, sizeof(JsonValue));
   value->value.object.elements = RedisModule_Calloc(size, sizeof(JsonKeyVal*));
   value->value.object.size = size;
